repair(tests): split logger tests and simplify mocks
- Split 'tests/unit/utils/logger.test.ts' into separate client and server test files.
- Created 'tests/unit/utils/logger.server.test.ts' for server-side logger tests.
- Simplified pino mock in server tests by removing unused 'levels' property.
- Ensured client tests only cover client-side logic without unnecessary server mocks.
